(refactor) dedicated instance folder
